Move StepButtons to own namespace.
