+++ osu.Framework.Testing/Drawables/StepButtons/AssertButton.cs
@@ -1,0 +1,29 @@
+// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
+
+using System;
+using OpenTK.Graphics;
+
+namespace osu.Framework.Testing.Drawables.StepButtons
+{
+    public class AssertButton : StepButton
+    {
+        public Func<bool> Assertion;
+
+        public string ExtendedDescription;
+
+        public AssertButton()
+        {
+            BackgroundColour = Color4.OrangeRed;
+            Action += checkAssert;
+        }
+
+        private void checkAssert()
+        {
+            if (Assertion())
+                BackgroundColour = Color4.YellowGreen;
+            else
+                throw new Exception($"{Text} {ExtendedDescription}");
+        }
+    }
+}
