+++ osu.Framework.Testing/osu.Framework.Testing.csproj
@@ -45,7 +45,10 @@
     <Reference Include="System.Core" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Drawables\StepButtons\AssertButton.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Drawables\StepButtons\RepeatStepButton.cs" />
+    <Compile Include="Drawables\StepButtons\StepButton.cs" />
     <Compile Include="TestBrowser.cs" />
     <Compile Include="TestBrowserConfig.cs" />
     <Compile Include="TestCase.cs" />
@@ -56,6 +59,7 @@
       <Project>{C76BF5B3-985E-4D39-95FE-97C9C879B83A}</Project>
       <Name>osu.Framework</Name>
     </ProjectReference>
+    <Compile Include="Drawables\StepButtons\ToggleStepButton.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\osu-framework.licenseheader">
